backend/requirements.txt
asgiref==3.11.1
Django==6.0.2
django-cors-headers==4.9.0
django-environ==0.12.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
python-dotenv==1.0.1
djangorestframework-camel-case==1.4.2
PyJWT==2.11.0
sqlparse==0.5.5
django-filter==24.3
pydantic==2.10.5
bleach==6.2.0
psycopg2-binary==2.9.9
redis==5.0.1
django-redis==5.4.0
bandit==1.7.6
safety==2.3.5
